feat: REGAPIC initial implementation
This includes:
1) Proper GRPC transcoding
2) Mimimal REST Operations (go make it compilable)
3) Rely on resource names for tests values generation (to not fail http url path validation)
4) Empty Server Streaming method implementation
