+++ src/main/java/com/google/api/generator/gapic/composer/common/AbstractServiceClientTestClassComposer.java
@@ -470,7 +470,7 @@ public abstract class AbstractServiceClientTestClassComposer implements ClassCom
             VariableExpr.withVariable(
                 Variable.builder().setType(methodArg.type()).setName(methodArgName).build());
         argExprs.add(varExpr);
-        Expr valExpr = DefaultValueComposer.createDefaultValue(methodArg, resourceNames);
+        Expr valExpr = createDefaultValue(methodArg, resourceNames);
         methodExprs.add(
             AssignmentExpr.builder()
                 .setVariableExpr(varExpr.toBuilder().setIsDecl(true).build())
@@ -536,7 +536,8 @@ public abstract class AbstractServiceClientTestClassComposer implements ClassCom
                       TypeNode.withReference(
                           ConcreteReference.builder()
                               .setClazz(List.class)
-                              .setGenerics(Arrays.asList(repeatedPagedResultsField.type().reference()))
+                              .setGenerics(
+                                  Arrays.asList(repeatedPagedResultsField.type().reference()))
                               .build()))
                   .setName("resources")
                   .build());
@@ -734,6 +735,9 @@ public abstract class AbstractServiceClientTestClassComposer implements ClassCom
       Map<String, ResourceName> resourceNames,
       Map<String, Message> messageTypes);
 
+  protected abstract Expr createDefaultValue(
+      MethodArgument methodArg, Map<String, ResourceName> resourceNames);
+
   protected List<Statement> createRpcExceptionTestStatements(
       Method method,
       List<MethodArgument> methodSignature,
@@ -765,7 +769,7 @@ public abstract class AbstractServiceClientTestClassComposer implements ClassCom
             VariableExpr.withVariable(
                 Variable.builder().setType(methodArg.type()).setName(methodArgName).build());
         argVarExprs.add(varExpr);
-        Expr valExpr = DefaultValueComposer.createDefaultValue(methodArg, resourceNames);
+        Expr valExpr = createDefaultValue(methodArg, resourceNames);
         tryBodyExprs.add(
             AssignmentExpr.builder()
                 .setVariableExpr(varExpr.toBuilder().setIsDecl(true).build())
