+++ src/main/java/com/google/api/generator/gapic/composer/grpc/ServiceClientTestClassComposer.java
@@ -1055,6 +1055,12 @@ public class ServiceClientTestClassComposer extends AbstractServiceClientTestCla
   }
 
   @Override
+  protected Expr createDefaultValue(
+      MethodArgument methodArg, Map<String, ResourceName> resourceNames) {
+    return DefaultValueComposer.createDefaultValue(methodArg, resourceNames, false);
+  }
+
+  @Override
   protected List<Statement> createRpcLroExceptionTestCatchBody(
       VariableExpr exceptionExpr, boolean isStreaming) {
     List<Expr> catchBodyExprs = new ArrayList<>();
