+++ src/main/java/com/google/api/generator/gapic/composer/samplecode/ServiceClientSampleCodeComposer.java
@@ -1302,7 +1302,7 @@ public class ServiceClientSampleCodeComposer {
         .map(
             arg ->
                 !isStringTypedResourceName(arg, resourceNames)
-                    ? DefaultValueComposer.createDefaultValue(arg, resourceNames)
+                    ? DefaultValueComposer.createDefaultValue(arg, resourceNames, false)
                     : stringResourceNameDefaultValueExpr.apply(arg))
         .collect(Collectors.toList());
   }
