+++ src/test/java/com/google/api/generator/gapic/composer/common/BUILD.bazel
@@ -5,8 +5,6 @@ package(default_visibility = ["//visibility:public"])
 
 UPDATE_GOLDENS_TESTS = [
     "BatchingDescriptorComposerTest",
-    "ServiceClientClassComposerTest",
-    "ServiceStubClassComposerTest",
 ]
 
 TESTS = UPDATE_GOLDENS_TESTS + [
