+++ src/test/java/com/google/api/generator/gapic/composer/grpc/BUILD.bazel
@@ -8,8 +8,10 @@ TESTS = [
     "GrpcServiceStubClassComposerTest",
     "MockServiceClassComposerTest",
     "MockServiceImplClassComposerTest",
+    "ServiceClientClassComposerTest",
     "ServiceClientTestClassComposerTest",
     "ServiceSettingsClassComposerTest",
+    "ServiceStubClassComposerTest",
     "ServiceStubSettingsClassComposerTest",
 ]
