+++ src/test/java/com/google/api/generator/gapic/composer/grpc/goldens/BookshopClient.golden
