+++ src/test/java/com/google/api/generator/gapic/composer/rest/goldens/HttpJsonComplianceStub.golden
@@ -12,6 +12,7 @@ import com.google.api.gax.httpjson.HttpJsonStubCallableFactory;
 import com.google.api.gax.httpjson.ProtoMessageRequestFormatter;
 import com.google.api.gax.httpjson.ProtoMessageResponseParser;
 import com.google.api.gax.httpjson.ProtoRestSerializer;
+import com.google.api.gax.httpjson.longrunning.stub.HttpJsonOperationsStub;
 import com.google.api.gax.rpc.ClientContext;
 import com.google.api.gax.rpc.UnaryCallable;
 import com.google.showcase.v1beta1.RepeatRequest;
@@ -67,7 +68,7 @@ public class HttpJsonComplianceStub extends ComplianceStub {
                             @Override
                             public String extract(RepeatRequest request) {
                               return ProtoRestSerializer.create()
-                                  .toBody("serverVerify", request.getServerVerify());
+                                  .toBody("*", request.toBuilder().build());
                             }
                           })
                       .build())
@@ -343,6 +344,7 @@ public class HttpJsonComplianceStub extends ComplianceStub {
   private final UnaryCallable<RepeatRequest, RepeatResponse> repeatDataPathTrailingResourceCallable;
 
   private final BackgroundResource backgroundResources;
+  private final HttpJsonOperationsStub operationsStub;
   private final HttpJsonStubCallableFactory callableFactory;
 
   public static final HttpJsonComplianceStub create(ComplianceStubSettings settings)
@@ -382,6 +384,7 @@ public class HttpJsonComplianceStub extends ComplianceStub {
       HttpJsonStubCallableFactory callableFactory)
       throws IOException {
     this.callableFactory = callableFactory;
+    this.operationsStub = HttpJsonOperationsStub.create(clientContext, callableFactory);
 
     HttpJsonCallSettings<RepeatRequest, RepeatResponse> repeatDataBodyTransportSettings =
         HttpJsonCallSettings.<RepeatRequest, RepeatResponse>newBuilder()
@@ -450,6 +453,10 @@ public class HttpJsonComplianceStub extends ComplianceStub {
     methodDescriptors.add(repeatDataPathResourceMethodDescriptor);
     methodDescriptors.add(repeatDataPathTrailingResourceMethodDescriptor);
     return methodDescriptors;
+  }
+
+  public HttpJsonOperationsStub getOperationsStub() {
+    return operationsStub;
   }
 
   @Override
