+++ test/integration/goldens/compute/com/google/cloud/compute/v1/stub/HttpJsonRegionOperationsStub.java
@@ -28,6 +28,7 @@ import com.google.api.gax.httpjson.HttpJsonStubCallableFactory;
 import com.google.api.gax.httpjson.ProtoMessageRequestFormatter;
 import com.google.api.gax.httpjson.ProtoMessageResponseParser;
 import com.google.api.gax.httpjson.ProtoRestSerializer;
+import com.google.api.gax.httpjson.longrunning.stub.HttpJsonOperationsStub;
 import com.google.api.gax.rpc.ClientContext;
 import com.google.api.gax.rpc.UnaryCallable;
 import com.google.cloud.compute.v1.GetRegionOperationRequest;
@@ -99,6 +100,7 @@ public class HttpJsonRegionOperationsStub extends RegionOperationsStub {
   private final UnaryCallable<GetRegionOperationRequest, Operation> getCallable;
 
   private final BackgroundResource backgroundResources;
+  private final HttpJsonOperationsStub operationsStub;
   private final HttpJsonStubCallableFactory callableFactory;
 
   public static final HttpJsonRegionOperationsStub create(RegionOperationsStubSettings settings)
@@ -139,6 +141,7 @@ public class HttpJsonRegionOperationsStub extends RegionOperationsStub {
       HttpJsonStubCallableFactory callableFactory)
       throws IOException {
     this.callableFactory = callableFactory;
+    this.operationsStub = HttpJsonOperationsStub.create(clientContext, callableFactory);
 
     HttpJsonCallSettings<GetRegionOperationRequest, Operation> getTransportSettings =
         HttpJsonCallSettings.<GetRegionOperationRequest, Operation>newBuilder()
@@ -158,6 +161,10 @@ public class HttpJsonRegionOperationsStub extends RegionOperationsStub {
     List<ApiMethodDescriptor> methodDescriptors = new ArrayList<>();
     methodDescriptors.add(getMethodDescriptor);
     return methodDescriptors;
+  }
+
+  public HttpJsonOperationsStub getOperationsStub() {
+    return operationsStub;
   }
 
   @Override
